refactor project commits response

+++ middleware_loader/core/domain/dtos/response/project_dto.go
@@ -65,12 +65,12 @@ func (in *GithubRepoResponseDTO) MapperListToGraphQLModel(input []GithubRepoResp
 }
 
 type ProjectCommitResponseDTO struct {
-	Id            string `json:"id"`
-	ProjectId     string `json:"projectId"`
-	ProjectName   string `json:"projectName"`
-	GithubRepo    string `json:"githubRepo"`
-	GithubRepoUrl string `json:"githubRepoUrl"`
-	UserCommitId  int    `json:"userCommitId"`
+	Id            string  `json:"id"`
+	ProjectId     string  `json:"projectId"`
+	ProjectName   string  `json:"projectName"`
+	GithubRepo    string  `json:"githubRepo"`
+	GithubRepoUrl string  `json:"githubRepoUrl"`
+	UserCommitId  float64 `json:"userCommitId"`
 }
 
 func NewProjectCommitResponseDTO() *ProjectCommitResponseDTO {
